Added order confirmation functions.

+++ store/models.py
@@ -12,7 +12,7 @@ class Product(models.Model):
     productName     = models.CharField(max_length=200, unique=True)
     slug            = models.SlugField(max_length=200, unique=True)
     description     = models.CharField(max_length=500, blank=True)
-    price           = models.IntegerField()
+    price           = models.DecimalField(max_digits=8, decimal_places=2)
     images          = models.ImageField(upload_to='photos/products')
     stock           = models.IntegerField()
     isAvailable     = models.BooleanField(default=True)
